Refactor: Refactoring the DTOs file by separating each DTO into a separated file.
