+++ Controllers/FreelancersController.cs
@@ -1,6 +1,6 @@
 ﻿using AonFreelancing.Contexts;
 using AonFreelancing.Models;
-using AonFreelancing.Models.DTOs;
+using AonFreelancing.Models.DTOs.FreelancerDTOs;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
