--- Models/DTOs/FreelancerDTO.cs
@@ -1,8 +1,0 @@
-﻿namespace AonFreelancing.Models.DTOs
-{
-    public class FreelancerDTO:UserDTO
-    {
-
-        public string Skills { get; set; }
-    }
-}
