+++ Models/DTOs/FreelancerDTOs/FreelancerDTO.cs
@@ -1,0 +1,10 @@
+﻿using AonFreelancing.Models.DTOs.UserDTOs;
+
+namespace AonFreelancing.Models.DTOs.FreelancerDTOs
+{
+    public class FreelancerDTO : UserDTO
+    {
+
+        public string Skills { get; set; }
+    }
+}
