+++ Models/DTOs/FreelancerDTOs/FreelancerInputDTO.cs
@@ -1,7 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.Diagnostics.CodeAnalysis;
+using AonFreelancing.Models.DTOs.UserDTOs;
 
-namespace AonFreelancing.Models.DTOs
+namespace AonFreelancing.Models.DTOs.FreelancerDTOs
 {
     public class FreelancerInputDTO : UserDTO
     {
