+++ Models/DTOs/UserDTOs/UserDTO.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
-namespace AonFreelancing.Models.DTOs
+namespace AonFreelancing.Models.DTOs.UserDTOs
 {
     public class UserDTO
     {
@@ -11,7 +11,7 @@ namespace AonFreelancing.Models.DTOs
         [StringLength(32)]
         public string Username { get; set; }
 
-        [MinLength(4,ErrorMessage ="Too short password")]
+        [MinLength(4, ErrorMessage = "Too short password")]
         public string Password { get; set; }
